Create a test for createSquire function and organize the project
