+++ src/types.ts
@@ -1,7 +1,3 @@
-import { Fighter } from "./characters/fighter/type";
-
-export type FidelityRange = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10;
-
 export interface mandatoryData {
   name: string;
   surname?: string;
@@ -12,10 +8,4 @@ export interface Character extends mandatoryData {
   isAlive: boolean;
   die: () => void;
   talk: () => string;
-}
-
-export interface Squire extends Character {
-  servedCharacter: Fighter;
-  fidelity: FidelityRange;
-  talk: () => "Soy un loser";
 }
